+++ docs/changelog.md
@@ -4,9 +4,13 @@ CHANGELOG
 0.7.1 (unreleased)
 ------------------
 
+**🚀 Nouveautés**
+* La gestion des permissions est définie pour chaque objet (module, site, visite) et l'objet ALL n'est plus pris en compte (#249). De fait les paramètres cruved des fichiers de configuration ainsi que permission object de `module.json` sont obsolètes.
+
 **🐛 Corrections**
 
 * export avec un filtre par jeux de données (#241)
+
 
 0.7.0 (2023-08-23)
 ------------------